(Clean Arch) - Finalização dos ajustes para novo padrão de arquitetura (#3)
* refactor(general): Realizando ajustes do padrão do projeto.

* refactor(IT): Utilizando testcontainers para testes de integração.

* refactor(Auth): Ajustes na classe de response de auth.

* refactor(WorkOrder): Liberando rota de listagem de workOrders

* refactor(buildAndVariables): Ajustando docker para criar e utilizar usuário não root, e configurando variaveis de ambiente

* refactor(infra): Realizando ajustes no pacote de infra

* refactor(pipe-owasp): Configurando .env

* refactor(sonar): Ajustes sonar

* refactor(sonar): Ajustes sonar

* refactor(sonar): Ajustes sonar

* refactor(tests): Criando testes.

* refactor(workorder,general): Refatorando package de work order e outros ajustes de padronização

* refactor(general): Resolvendo smells

* refactor(general): Resolvendo smells

* refactor(general): Resolvendo smells

* refactor(general): Ajustes de padronizacao

* refactor(messages): Adicionando traduções faltantes

* refactor(workOrder): Ajustando retorno.

* refactor(workOrder): Ajustando mapeamento de classes

* refactor(workOrder): Adicionando tradução faltante

* refactor(workOrder): Removendo tradução não utilizadas

+++ src/main/java/com/mecanicadm/mecanicadm_api/infra/features/user/persistence/jpa/PasswordResetTokenRepositoryImpl.java
@@ -3,9 +3,13 @@ package com.mecanicadm.mecanicadm_api.infra.features.user.persistence.jpa;
 import com.mecanicadm.mecanicadm_api.core.user.domain.PasswordResetToken;
 import com.mecanicadm.mecanicadm_api.core.user.domain.User;
 import com.mecanicadm.mecanicadm_api.core.user.domain.port.PasswordResetTokenGateway;
+import com.mecanicadm.mecanicadm_api.shared.exception.TechnicalException;
 import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
+
+import static java.util.Objects.isNull;
 
 @Repository
 public class PasswordResetTokenRepositoryImpl implements PasswordResetTokenGateway {
@@ -17,12 +21,19 @@ public class PasswordResetTokenRepositoryImpl implements PasswordResetTokenGatew
     }
 
     @Override
+    @Transactional
     public void deleteByUser(User user) {
+        if (isNull(user)) {
+            throw new TechnicalException("error.technical.entity.null", "User", "exclusão");
+        }
         jpaRepository.deleteByUser(UserJpaMapper.toEntity(user));
     }
 
     @Override
     public void save(PasswordResetToken token) {
+        if (isNull(token)) {
+            throw new TechnicalException("error.technical.entity.null", "PasswordResetToken", "salvar");
+        }
         jpaRepository.save(PasswordResetTokenJpaMapper.toEntity(token));
     }
 
@@ -33,6 +44,9 @@ public class PasswordResetTokenRepositoryImpl implements PasswordResetTokenGatew
 
     @Override
     public void delete(PasswordResetToken token) {
+        if (isNull(token)) {
+            throw new TechnicalException("error.technical.entity.null", "PasswordResetToken", "exclusão");
+        }
         jpaRepository.delete(PasswordResetTokenJpaMapper.toEntity(token));
     }
 }
